add functionality for adding/creating user event from Explore page
